# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Telephone-Booth Operator API
  description: |
    Versioned REST + WebSocket API between the operator console
    (`Telephone-Booth-Operator`) and the phone-side Rust client.

    The phone client authenticates with a static Bearer API token issued from
    the operator UI. Operator browser sessions authenticate via Authentik OIDC
    (cookie session). Native mobile apps in
    `Telephone-Booth-Operator-Mobile` authenticate with Authentik PKCE and send
    the resulting access token as `Authorization: Bearer <jwt>`
    (`operatorBearer` security scheme).
  version: 1.0.0
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8787
    description: Local dev
  - url: https://operator.example.com
    description: Production
...
paths:
...
  /v1/messages:
    get:
      summary: List messages, filterable by status and creation time.
      tags: [messages]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - name: installationId
          in: query
          description: |
            Installation scope. Omit for the currently active installation
            (the default), pass a uuid for a historical one, or `all` to span
            every installation.
          schema:
            oneOf:
              - { type: string, format: uuid }
              - { type: string, enum: [all] }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/MessageStatus" }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageList" }
    post:
      summary: Initiate a message audio upload.
      tags: [messages]
      security:
        - apiToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageCreate" }
      responses:
        "201":
          description: >-
            Upload URL issued. Also returned when re-initiating an existing
            message that is still `uploading` (e.g. the booth retried after a
            reboot): the same message id and a fresh SAS are handed back so the
            upload can resume.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageInitiated" }
        "404": { description: question not found }
        "409":
          description: >-
            Conflicting audio hash: the matching message has already advanced
            past `uploading`, the retry targets a different question, or the
            hash belongs to a non-message file.
  /v1/messages/random:
    get:
      summary: A random approved message for the phone client to play.
      tags: [messages]
      security:
        - apiToken: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "404": { description: no approved messages available }
  /v1/messages/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Fetch one message.
      tags: [messages]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "404": { description: not found }
    delete:
      summary: Delete a message row.
      tags: [messages]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "204": { description: deleted }
        "404": { description: not found }
  /v1/messages/{id}/complete:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Phone client notifies that the direct blob upload finished.
      description: |
        Verifies the uploaded blob and moves the message from `uploading`
        straight to `pending`, i.e. into the operator review queue.
        Transcription is optional enrichment pushed in later by the external
        Transcription app, so it never gates whether a landed recording is
        reviewable.
      tags: [messages]
      security:
        - apiToken: []
      responses:
        "200":
          description: upload accepted; message queued for review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageComplete" }
        "404": { description: not found }
        "409": { description: blob not found }
        "413": { description: audio blob exceeds maximum allowed size }
        "422": { description: "sha256 metadata missing or mismatch" }
  /v1/messages/{id}/transcriptions:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: List all transcription attempts for a message (newest first).
      tags: [transcriptions]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranscriptionList" }
        "404": { description: message not found }
  /v1/messages/{id}/transcribe:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Re-run transcription (and downstream moderation) for a message.
      description: |
        Runs the full pipeline synchronously — transcription, then
        moderation against the new transcript, then any configured
        auto-decision — and returns the resulting `Transcription` row.
        The same row (plus the matching `Message` update) is also
        broadcast on `/v1/ws/status` as a `{kind:"message"}` envelope so
        other clients pick up the change without polling. The status is
        `202` to signal that the row may still be of interest to other
        listeners (history queries, the moderation row, the parent
        `Message`), even though the synchronous work is finished.
      tags: [transcriptions]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "202":
          description: transcription attempt recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transcription" }
        "404": { description: message not found }
  /v1/messages/{id}/transcription:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Submit an operator-authored transcript for a message.
      description: |
        Lets a logged-in operator record transcript text directly — for
        example the iOS Transcriber app doing on-device transcription — rather
        than triggering the server-side pipeline via `/transcribe`. Mirrors the
        worker push-back callback semantics: a pending transcription row is
        finalized, otherwise a new succeeded row is recorded (a redelivery of
        the latest succeeded transcript — same `text`, `language`, `model` and
        submitter — is a no-op, while a correction records a new row). The row is
        attributed to the submitting operator (`requestedById`). Like all
        pushed transcripts the text is still translated and moderated
        server-side before it can be acted on. `text` may be empty for a silent
        recording. The resulting `Transcription` row is returned and broadcast
        on `/v1/ws/status` as a `{kind:"message"}` envelope.
      tags: [transcriptions]
      security:
        - operatorSession: []
        - operatorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 20000 }
                language: { type: [string, "null"], minLength: 1, maxLength: 64 }
                model: { type: [string, "null"], minLength: 1, maxLength: 128 }
      responses:
        "202":
          description: transcription recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transcription" }
        "404": { description: message not found }
  /v1/messages/{id}/moderate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Re-run moderation against the latest succeeded transcription.
      description: |
        Runs the moderation step (and any configured auto-decision)
        synchronously against the latest succeeded transcription and
        returns the resulting `Moderation` row. The same row is also
        broadcast on `/v1/ws/status` as a `{kind:"message"}` envelope.
      tags: [moderations]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "202":
          description: moderation attempt recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Moderation" }
        "404": { description: message not found }
        "409": { description: no succeeded transcription available to moderate }
  /v1/messages/{id}/decision:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Record a human moderation decision (approve or reject).
      description: |
        Lets a logged-in operator override (or stand in for) the AI
        pipeline by approving or rejecting a message. Sets the message
        `status`, `decidedAt`, and `decidedById`, and optionally replaces
        `notes`. Only valid once the recording has landed — a message that
        is still `uploading` returns `409`.
      tags: [messages]
      security:
        - operatorSession: []
        - operatorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approve, reject] }
                notes: { type: string, maxLength: 2000 }
      responses:
        "200":
          description: decision recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "404": { description: message not found }
        "409": { description: message is still uploading and cannot be decided }
  /v1/messages/{id}/translation:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Attach a human translation to the latest succeeded transcription.
      description: |
        Lets a logged-in operator supply an English translation for a
        transcription the translation worker could not produce. Updates the
        message's latest succeeded transcription, marking `translationStatus`
        as `succeeded` with a null `translationProvider` (consumers infer a
        manual translation from a succeeded translation with no provider).
        Returns `409` when the message has no succeeded transcription yet.
      tags: [messages]
      security:
        - operatorSession: []
        - operatorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [translatedText]
              properties:
                translatedText: { type: string, minLength: 1, maxLength: 20000 }
                translatedLanguage: { type: string, minLength: 1, maxLength: 64 }
      responses:
        "200":
          description: translation recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transcription" }
        "404": { description: message not found }
        "409": { description: no succeeded transcription to translate }
...
  /v1/worker/messages/{id}/transcription:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Push-mode transcription result callback.
      description: |
        The Transcription app (macOS + iOS) transcribes booth audio and POSTs
        the text back here. Transcription is optional: a completed message is
        already `pending` in the operator review queue, so the app decides on
        its own when (or whether) to transcribe and the server never solicits
        the work. Unsolicited results are accepted — if no pending
        transcription row exists the server records a new succeeded one.
        Authenticated with a static operator-issued API token (same scheme as
        phone clients). Callbacks are claim-free (no leases); the server guards
        every write so a stale/duplicate callback cannot downgrade a finalized
        row. On success the server creates the pending moderation row and
        broadcasts the next `work` step (`translation` for non-English audio,
        else `moderation`).
      tags: [worker]
      security:
        - apiToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerTranscriptionBody" }
      responses:
        "200": { description: Result accepted. }
        "404": { description: Message not found. }
  /v1/worker/messages/{id}/translation:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Push-mode translation result callback.
      description: |
        Pushed back after the worker translates a non-English transcript to
        English (triggered by a `{kind:"work", needs:["translation"]}`
        envelope). On success the server broadcasts the `moderation` work step.
      tags: [worker]
      security:
        - apiToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerTranslationBody" }
      responses:
        "200": { description: Result accepted. }
        "404": { description: Message or transcription not found. }
...
components:
  securitySchemes:
    apiToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Phone-client API token issued from the operator UI.
    operatorSession:
      type: apiKey
      in: cookie
      name: __Host-booth_session
      description: HMAC-signed opaque session cookie set after OIDC login.
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: jwt
      description: |
        Authentik-issued JWT access token. Validated against the provider's
        JWKS, issuer, audience (operator web client_id + any configured
        `OIDC_MOBILE_AUDIENCES`), and operator group / email allow-lists.
        Used by native mobile apps that authenticate via PKCE.
  schemas:
...
    Sha256:
      type: string
      pattern: "^[a-f0-9]{64}$"
...
    MessageStatus:
      type: string
      enum: [uploading, received, pending, approved, rejected]
    Message:
      type: object
      required: [id, status, createdAt, audio]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/MessageStatus" }
        questionId: { anyOf: [{ type: string, format: uuid }, { type: "null" }] }
        notes: { anyOf: [{ type: string }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        receivedAt: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
        decidedAt: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
        decidedById: { anyOf: [{ type: string }, { type: "null" }] }
        audio: { $ref: "#/components/schemas/AudioRef" }
        latestTranscription:
          anyOf:
            - $ref: "#/components/schemas/Transcription"
            - { type: "null" }
        latestModeration:
          anyOf:
            - $ref: "#/components/schemas/Moderation"
            - { type: "null" }
    MessageCreate:
      type: object
      required: [durationMs, sha256]
      properties:
        questionId: { type: string, format: uuid }
        durationMs: { type: integer, minimum: 1, maximum: 300000 }
        sha256: { $ref: "#/components/schemas/Sha256" }
    MessageInitiated:
      type: object
      required: [id, uploadUrl, blobName]
      properties:
        id: { type: string, format: uuid }
        uploadUrl: { type: string, format: uri }
        blobName: { type: string }
    MessageComplete:
      type: object
      required: [id, status, receivedAt]
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [uploading, received, pending, approved, rejected]
          description: >
            Status after the transition. A newly completed upload is
            `pending`; other values only occur on an idempotent replay of a
            message that already advanced.
        receivedAt: { type: string, format: date-time }
    MessageList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Message" }
...
    Transcription:
      type: object
      required: [id, messageId, provider, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        messageId: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/AiProvider" }
        model: { anyOf: [{ type: string }, { type: "null" }] }
        status: { $ref: "#/components/schemas/TranscriptionStatus" }
        text: { anyOf: [{ type: string }, { type: "null" }] }
        language: { anyOf: [{ type: string }, { type: "null" }] }
        durationMs: { anyOf: [{ type: integer, minimum: 0 }, { type: "null" }] }
        latencyMs: { anyOf: [{ type: integer, minimum: 0 }, { type: "null" }] }
        error: { anyOf: [{ type: string }, { type: "null" }] }
        requestedById: { anyOf: [{ type: string }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        completedAt: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
...
    TranscriptionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Transcription" }
...